Move RootPath to fs package
This moves the RootPath function out of the archive package and into the
fs package for external use.
